+++ types/errors/abci.go
@@ -121,16 +121,15 @@ func debugErrEncoder(err error) string {
 	return fmt.Sprintf("%+v", err)
 }
 
-// The defaultErrEncoder applies Redact on the error before encoding it with its internal error message.
 func defaultErrEncoder(err error) string {
-	return Redact(err).Error()
+	return err.Error()
 }
 
 type coder interface {
 	ABCICode() uint32
 }
 
-// abciCode test if given error contains an ABCI code and returns the value of
+// abciCode tests if given error contains an ABCI code and returns the value of
 // it if available. This function is testing for the causer interface as well
 // and unwraps the error.
 func abciCode(err error) uint32 {
@@ -189,20 +188,4 @@ func errIsNil(err error) bool {
 		return val.IsNil()
 	}
 	return false
-}
-
-var errPanicWithMsg = Wrapf(ErrPanic, "error message redacted to hide potential sensitive info. Use the '--trace' flag if you are running a node to see the full stack trace error")
-
-// Redact replaces an error that is not initialized as an ABCI Error with a
-// generic internal error instance. If the error is an ABCI Error, that error is
-// simply returned.
-func Redact(err error) error {
-	if ErrPanic.Is(err) {
-		return errPanicWithMsg
-	}
-	if abciCode(err) == internalABCICode {
-		return errInternal
-	}
-
-	return err
 }